requirements.txt
easydict==1.13
fair_esm==2.0.0
Jinja2==3.1.4
numpy==1.24.3
numpy==1.23.3
pandas==1.5.0
pandas==1.5.3
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.1.2
torch==1.12.0+cu113
torchdrug==0.2.1
xgboost==1.6.2
